Shortened-Url/requirements.txt
Django==1.10.5
argparse==1.2.1
djangorestframework==3.5.3
wsgiref==0.1.2
requests==2.13.0